Fixed movability of the handler
Due to the use of the direct pointers to the handler within the registry
the move operation of the handler was broken. To solve the issue, a
unique pointer is used that holds an instance of the opaque handle
structure.

Another fact is that it is perfectly legal to destroy a signal handler
during its callback execution. The thread callback executing thread is
not waited for in this case, the deletion only causes the handler to
deregister itself.